+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 ## [Unrelease]
 ### [Added]
+- Add the head file `includes/LoggingCpp.h` to define the namespace and class.
 
 ### [Changed]